feat(dh): rename pipes
