+++ libs/dh/metering-point/feature-overview/src/lib/breadcrumb/dh-breadcrumb.component.html
@@ -20,7 +20,7 @@ limitations under the License.
   }}</a>
   <watt-icon name="right"></watt-icon>
   <ng-container
-    *ngIf="meteringPoint?.meteringPointType | isParent; else isChild"
+    *ngIf="meteringPoint?.meteringPointType | dhIsParent; else isChild"
   >
     {{ meteringPoint?.gsrnNumber }}
   </ng-container>
