+++ libs/dh/metering-point/feature-overview/src/lib/dh-metering-point-overview.component.html
@@ -38,7 +38,7 @@ limitations under the License.
         <watt-tab [label]="t('prices')">
           <dh-charges></dh-charges>
         </watt-tab>
-        <ng-container *ngIf="meteringPoint?.meteringPointType | isParent">
+        <ng-container *ngIf="meteringPoint?.meteringPointType | dhIsParent">
           <watt-tab
             [label]="
               t('childMeteringPoints', {
