+++ libs/dh/metering-point/feature-overview/src/lib/identity/dh-metering-point-identity.component.html
@@ -20,11 +20,11 @@ limitations under the License.
 >
   <div class="watt-text-s">
     <span [title]="t('meteringPointType')">{{
-      translationKeys?.meteringPointType | transloco | emDashFallback
+      translationKeys?.meteringPointType | transloco | dhEmDashFallback
     }}</span>
 
     (<span [title]="t('meteringMethod')">{{
-      translationKeys?.meteringMethod | transloco | emDashFallback
+      translationKeys?.meteringMethod | transloco | dhEmDashFallback
     }}</span
     >)
   </div>
