+++ libs/dh/metering-point/feature-overview/src/lib/primary-master-data/dh-metering-point-primary-master-data.component.ts
@@ -36,8 +36,8 @@ import {
 } from '@energinet-datahub/watt';
 
 import { DhEmDashFallbackPipeScam } from '../shared/dh-em-dash-fallback.pipe';
-import { DhYesNoPipeScam } from '../shared/yes-no.pipe';
-import { DhIsParentPipeScam } from '../shared/is-parent.pipe';
+import { DhYesNoPipeScam } from '../shared/dh-yes-no.pipe';
+import { DhIsParentPipeScam } from '../shared/dh-is-parent.pipe';
 import { DhShowForMeteringPointTypeDirectiveScam } from '../shared/dh-show-for-metering-point-type.directive';
 
 export type PrimaryMasterData = Pick<
