+++ libs/dh/metering-point/feature-overview/src/lib/secondary-master-data/dh-secondary-master-data.component.html
@@ -34,7 +34,7 @@ limitations under the License.
           {{ t('netSettlementGroup') }}
         </dt>
         <dd class="watt-text-m">
-          {{ netSettlementGroupAsNumber | emDashFallback }}
+          {{ netSettlementGroupAsNumber | dhEmDashFallback }}
         </dd>
       </div>
       <div
@@ -48,7 +48,9 @@ limitations under the License.
           {{ t('disconnectionType') }}
         </dt>
         <dd class="watt-text-m" data-testid="disconnectionType">
-          {{ translationKeys?.disconnectionType | transloco | emDashFallback }}
+          {{
+            translationKeys?.disconnectionType | transloco | dhEmDashFallback
+          }}
         </dd>
       </div>
       <div
@@ -63,7 +65,9 @@ limitations under the License.
         </dt>
         <dd class="watt-text-m">
           {{
-            secondaryMasterData?.productionObligation | yesNo | emDashFallback
+            secondaryMasterData?.productionObligation
+              | dhYesNo
+              | dhEmDashFallback
           }}
         </dd>
       </div>
@@ -78,7 +82,7 @@ limitations under the License.
           {{ t('connectionType') }}
         </dt>
         <dd class="watt-text-m" data-testid="connectionType">
-          {{ translationKeys?.connectionType | transloco | emDashFallback }}
+          {{ translationKeys?.connectionType | transloco | dhEmDashFallback }}
         </dd>
       </div>
       <div
@@ -92,7 +96,7 @@ limitations under the License.
           {{ t('inAreaCode') }}
         </dt>
         <dd class="watt-text-m">
-          {{ secondaryMasterData?.inAreaCode | emDashFallback }}
+          {{ secondaryMasterData?.inAreaCode | dhEmDashFallback }}
         </dd>
       </div>
       <div
@@ -106,7 +110,7 @@ limitations under the License.
           {{ t('outAreaCode') }}
         </dt>
         <dd class="watt-text-m">
-          {{ secondaryMasterData?.outAreaCode | emDashFallback }}
+          {{ secondaryMasterData?.outAreaCode | dhEmDashFallback }}
         </dd>
       </div>
       <div
@@ -120,8 +124,8 @@ limitations under the License.
           {{ t('gridAreaName') }}
         </dt>
         <dd class="watt-text-m">
-          {{ secondaryMasterData?.gridAreaName | emDashFallback }} ({{
-            secondaryMasterData?.gridAreaCode | emDashFallback
+          {{ secondaryMasterData?.gridAreaName | dhEmDashFallback }} ({{
+            secondaryMasterData?.gridAreaCode | dhEmDashFallback
           }})
         </dd>
       </div>
@@ -139,7 +143,7 @@ limitations under the License.
           {{
             translationKeys?.capacityFormatted
               | transloco: { capacity: secondaryMasterData?.capacity }
-              | emDashFallback
+              | dhEmDashFallback
           }}
         </dd>
       </div>
@@ -154,7 +158,7 @@ limitations under the License.
           {{ t('assetType') }}
         </dt>
         <dd class="watt-text-m" data-testid="assetType">
-          {{ translationKeys?.assetType | transloco | emDashFallback }}
+          {{ translationKeys?.assetType | transloco | dhEmDashFallback }}
         </dd>
       </div>
       <div
@@ -171,7 +175,7 @@ limitations under the License.
           {{
             translationKeys?.ratedCapacityFormatted
               | transloco: { capacity: secondaryMasterData?.ratedCapacity }
-              | emDashFallback
+              | dhEmDashFallback
           }}
         </dd>
       </div>
@@ -189,7 +193,7 @@ limitations under the License.
           {{
             translationKeys?.ratedCurrentFormatted
               | transloco: { current: secondaryMasterData?.ratedCurrent }
-              | emDashFallback
+              | dhEmDashFallback
           }}
         </dd>
       </div>
@@ -204,7 +208,7 @@ limitations under the License.
           {{ t('powerPlantGsrnNumber') }}
         </dt>
         <dd class="watt-text-m">
-          {{ secondaryMasterData?.powerPlantGsrnNumber | emDashFallback }}
+          {{ secondaryMasterData?.powerPlantGsrnNumber | dhEmDashFallback }}
         </dd>
       </div>
       <div
@@ -250,7 +254,7 @@ limitations under the License.
         </dd>
       </div>
     </dl>
-    <ng-container *ngIf="secondaryMasterData?.meteringPointType | isParent">
+    <ng-container *ngIf="secondaryMasterData?.meteringPointType | dhIsParent">
       <div class="watt-text-l">
         {{ t('locationDetailTitle') }}
       </div>
@@ -266,7 +270,7 @@ limitations under the License.
             {{ t('streetCode') }}
           </dt>
           <dd class="watt-text-m">
-            {{ secondaryMasterData?.streetCode | emDashFallback }}
+            {{ secondaryMasterData?.streetCode | dhEmDashFallback }}
           </dd>
         </div>
         <div
@@ -280,7 +284,7 @@ limitations under the License.
             {{ t('municipalityCode') }}
           </dt>
           <dd class="watt-text-m">
-            {{ secondaryMasterData?.municipalityCode | emDashFallback }}
+            {{ secondaryMasterData?.municipalityCode | dhEmDashFallback }}
           </dd>
         </div>
         <div
