+++ libs/dh/metering-point/feature-overview/src/lib/shared/dh-em-dash-fallback.pipe.ts
@@ -19,7 +19,7 @@ import { NgModule, Pipe, PipeTransform } from '@angular/core';
 import { emDash } from './em-dash';
 
 export type TValue = string | number | undefined | null;
-export const pipeName = 'emDashFallback';
+export const pipeName = 'dhEmDashFallback';
 
 @Pipe({ name: pipeName })
 export class DhEmDashFallbackPipe implements PipeTransform {
